Configuration change and fixes for static land-cover maps
Configuration files now should no longer specify the bounds on any parameter. For Uniform and other priors that require bounds, these should be specified in the same priors file as for others. Following new support for dynamic land-cover, it was necessary to make some slight adjustments so that static land-cover maps still work.