UPDATED MOD16 algorithm to correct boundary-layer issue from User Guide
In the Collection 6.1 User Guide, there was an error where the atmospheric boundary-layer resistance (as a function of VPD), a ramp function, was set to its minimum value when VPD was low and its maximum value when VPD was high; this was in error, as the reverse is in line with the original algorithm and more realistic (i.e., atmospheric demand for water vapor should be low when VPD is low, therefore resistance should be high). It was necessary to keep the original (incorrect) calculation to be consistent with Collection 6.1.