REVERTED boundary-layer resistance calculation; verified canopy evaporation against PGE617
Getting the same results for canopy evaporation as PGE617, which is good. The "boundary-layer resistance", after much internal debate, is better understood as the surface resistance, and results from using VPD as a proxy for soil moisture. Therefore, Equation 26 in Mu et al. (2011) is indeed incorrect, although it makes sense in the context of aerodynamic resistance.